requirements.txt
appdirs==1.4.4
attrs==20.3.0
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==5.0.5
distlib==0.3.1
filelock==3.0.12
fire==0.4.0
frictionless==4.6.0
identify==2.2.2
idna==2.10
isodate==0.6.0
jsonschema==3.2.0
nodeenv==1.5.0
petl==1.7.4
pre-commit==2.12.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-slugify==4.0.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
shellingham==1.4.0
simpleeval==0.9.10
six==1.15.0
stringcase==1.2.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
typer==0.3.2
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.3
